refactor: split interprocedural array safety analysis
